+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Convert `settings.filter` to a keyword argument.
+
 ## [0.7.1] - 2025-07-17
 
 ### Added